+++ CPE200Lab1/CPE200Lab1/RPNCalculatorEngine.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace CPE200Lab1
 {
-    public class RPNCalculatorEngine : CalculatorEngine
+    public class RPNCalculatorEngine : TheCalculatorEngine
     {
         /// <summary>
         /// Check that is unary operator.
